+++ src/wordlist.c
@@ -126,7 +126,7 @@ int wordlist_append_from_file(wordlist_t *wordlist, char *filename)
 
 void wordlist_destroy(wordlist_t *wordlist)
 {
-	for (int i = 0; i < wordlist->length; ++i) {
+	for (size_t i = 0; i < wordlist->length; ++i) {
 		free(wordlist->words[i]);
 	}
 	free(wordlist->words);
